added user seeds and updated schema

+++ db/schema/01_users.sql
@@ -5,7 +5,7 @@ CREATE TABLE users (
   id SERIAL PRIMARY KEY NOT NULL,
   name VARCHAR(255) NOT NULL,
   email VARCHAR(255) NOT NULL,
-  phone INTEGER NOT NULL,
-  backup_phone INTEGER
-  password VARCHAR(255) NOT NULL,
+  phone BIGINT NOT NULL,
+  backup_phone BIGINT,
+  password VARCHAR(255) NOT NULL
 );
